requirements.txt
requests
beautifulsoup4
lxml
aiohttp
pytest-asyncio